+++ liblepton/scheme/unit-tests/t0205-page-parse-unterminated-attribute-list.scm
@@ -1,8 +1,8 @@
 ;; Test that unterminated attribute lists cause a parse error.
 
 (use-modules (unit-test)
+             (lepton attrib)
              (lepton object)
-             (geda attrib)
              ((geda page) #:renamer (symbol-prefix-proc 'geda:))
              (lepton page))