+++ liblepton/scheme/unit-tests/t0206-page-parse-garbage-attribute.scm
@@ -1,8 +1,8 @@
 ;; Test that garbage in attribute lists causes a parse error.
 
 (use-modules (unit-test)
+             (lepton attrib)
              (lepton object)
-             (geda attrib)
              ((geda page) #:renamer (symbol-prefix-proc 'geda:))
              (lepton page))