+++ netlist/scheme/netlist/schematic-connection.scm
@@ -23,7 +23,7 @@
   #:use-module (srfi srfi-9)
   #:use-module (srfi srfi-9 gnu)
   #:use-module (srfi srfi-26)
-  #:use-module (geda attrib)
+  #:use-module (lepton attrib)
   #:use-module (lepton object)
   #:use-module (lepton page)
   #:use-module (netlist package-pin)