+++ symcheck/scheme/symbol/check/obsolete.scm
@@ -17,7 +17,7 @@
 ;;; Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111 USA
 
 (define-module (symbol check obsolete)
-  #:use-module (geda attrib)
+  #:use-module (lepton attrib)
   #:use-module (lepton object)
   #:use-module (symbol gettext)
   #:use-module (symbol blame)