+++ symcheck/scheme/symbol/check/slot.scm
@@ -22,12 +22,12 @@
   #:use-module (srfi srfi-9)
   #:use-module (srfi srfi-9 gnu)
   #:use-module (srfi srfi-26)
+  #:use-module (lepton attrib)
+  #:use-module (lepton object)
   #:use-module (symbol gettext)
   #:use-module (symbol blame)
   #:use-module (symbol check attrib)
   #:use-module (symbol check duplicate)
-  #:use-module (geda attrib)
-  #:use-module (lepton object)
 
   #:export-syntax (make-slot slot?
                    slot-object set-slot-object!