+++ NEWS.md
@@ -20,6 +20,10 @@ Notable changes in Lepton EDA 1.9.10 (upcoming)
   among various build environments.
 
 ### Scheme API changes:
+- The module `(geda attrib)` has been renamed to `(lepton attrib)`.
+  All the functions and variables available are still re-exported
+  in the former one but its using is discouraged.
+
 - The module `(geda object)` has been renamed to `(lepton object)`.
   All the functions and variables available are still re-exported
   in the former one but its using is discouraged.